修复UI反馈问题

+++ proj/HBDJProj/HBDJProj/Classes/Discovery/subPart/subPartDetail/view/DCRichTextTopInfoView.m
@@ -22,7 +22,7 @@
 @implementation DCRichTextTopInfoView
 
 - (void)reloadPlayCount:(NSInteger)count{
-    _counts.text = [NSString stringWithFormat:@"查看次数: %ld",count];
+    _counts.text = [NSString stringWithFormat:@"查看次数: %ld",(long)count];
 }
 
 - (void)setModel:(DJDataBaseModel *)model{
